types: further improve Polymer interop types.
This change adds better typing pattern for tf-paginated-view and
vz-histograms.
